requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
python-multipart>=0.0.9,<1.0.0
pydantic>=2.8.0,<3.0.0
pydantic-settings>=2.4.0,<3.0.0
python-dotenv>=1.0.1,<2.0.0
langchain>=0.2.14,<1.0.0
langchain-core>=0.2.33,<1.0.0
langchain-groq>=0.1.9,<1.0.0
langchain-text-splitters>=0.2.2,<1.0.0
langchain-chroma>=0.1.2,<1.0.0
langchain-huggingface>=0.0.3,<1.0.0
chromadb>=0.5.5,<1.0.0
sentence-transformers>=3.0.1,<4.0.0
docling>=2.8.1,<3.0.0
pypdf>=4.3.1,<5.0.0
pillow>=10.4.0,<11.0.0
pytesseract>=0.3.10,<1.0.0
streamlit>=1.37.1,<2.0.0
requests>=2.32.3,<3.0.0
ragas>=0.1.14,<1.0.0
datasets>=2.20.0,<3.0.0
pandas>=2.2.2,<3.0.0
numpy>=1.26.4,<3.0.0
tenacity>=8.5.0,<9.0.0
pytest>=8.3.0,<9.0.0
